elevator and fsm (possibly) 100% functional
